test_end_to_end: separate into tests, data, and utils, and WIP description
